Tons of Experience Section work
Added in animation, not perfectly aligned, but will have to do for now.

Added in experience blocks, need to make it look better, but struggling due to it being within an SVG. Unsure if i want to pull it out and make it overlay.

todo:
Mobile compatibility by reusing the old experience section. Mobile can be more simple and to the point.

Refactor the code to have comments explaining, and take another look at the utility function for text wrapping due to the SVG properties.